% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cuda_provenance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cudatensr}{\code{\link[cudatensr:cuda_provenance]{cuda_provenance()}}}
}}

