% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{sparse_info}
\alias{sparse_info}
\title{Inspect sparse matrix metadata}
\usage{
sparse_info(x)
}
\arguments{
\item{x}{A \code{cudasparse} matrix.}
}
\value{
A named list.
}
\description{
Inspect sparse matrix metadata
}
\examples{
sparse_info(cuda_sparse(diag(3), device = "cpu"))
}
